FIX DUPLICATE STDOUT
