+++ source/Sailfish.TestAdapter/TestSettingsParser/SailfishSettings.cs
@@ -1,0 +1,11 @@
+using System.Text.Json.Serialization;
+
+namespace Sailfish.TestAdapter.TestSettingsParser;
+
+#pragma warning disable CS8618
+public class SailfishSettings
+{
+    [JsonPropertyName("_comment")] public string Comment { get; set; }
+
+    public SailfishTestSettings TestSettings { get; set; }
+}
